Refactor mergeCompactChunks into smaller functions
