+++ pkg/bloomcompactor/TODO.md
@@ -1,3 +1,2 @@
-* Use tarGz, untarGz before uploding blocks to storage
 * Introduce back `maxLookBackPeriod` as `RejectOldSamplesMaxAge` limit in distributors
 * restrict block size during creation. Suggestion to use cut logic we use for memchunks. Suggested start size is 500mb
