+++ pkg/bloomcompactor/chunkcompactor_test.go
@@ -187,7 +187,7 @@ func TestLazyBloomBuilder(t *testing.T) {
 	require.Equal(t, 6, mcc.chunkCount)
 	require.Equal(t, fp3, it.At().Series.Fingerprint)
 
-	// interator is done
+	// iterator is done
 	require.False(t, it.Next())
 	require.Error(t, io.EOF, it.Err())
 	require.Equal(t, v1.SeriesWithBloom{}, it.At())
